Extract infrastructure and explain architecture
