+++ packages/gitgraph-cli/src/render.ts
@@ -7,6 +7,11 @@ export interface IRenderGraph {
   branchOpen(): void;
 }
 
+// Translate rendered data into CLI logic.
+//
+// This is necessary because rendered data are optimized for browsers.
+// Rendering is a bit different in CLI since we don't have pixels.
+// Thus, we should translate data to have "line-per-line" instructions.
 function render(logger: IRenderGraph, { commits }: RenderedData): void {
   commits.forEach((commit, index) => {
     const remainingCommits = commits.slice(index + 1);
